requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
jinja2==3.1.4
python-multipart==0.0.9
passlib==1.7.4
pypdf==5.0.1
python-docx==1.1.2
itsdangerous==2.2.0
httpx==0.27.2
